// src/payload-types.ts
export interface User {
  id: number;
  name?: string | null;
  /**
   * Controls admin panel access and what this person can change.
   */
  roles: ('admin' | 'editor' | 'sales' | 'customer')[];
  updatedAt: string;
  createdAt: string;
  email: string;
  resetPasswordToken?: string | null;
  resetPasswordExpiration?: string | null;
  salt?: string | null;
  hash?: string | null;
  loginAttempts?: number | null;
  lockUntil?: string | null;
  sessions?:
    | {
        id: string;
        createdAt?: string | null;
        expiresAt: string;
      }[]
    | null;
  password?: string | null;
  collection: 'users';
}